Update jQuery to latest (1.8.3)
This change is not necessary to fix any bugs, but it is intended to
prevent confusion, particularly for users of `api.jquery.com` and
`learn.jquery.com`. These sites describe the latest version of jQuery,
so the globally-available jQuery should likewise be kept up-to-date.
This will prevent any confusion that might otherwise arise when users
experiment with jQuery from their browser's console.