Refactor the population of the WordPress current network object
As of WordPress 3.9, `wpmu_current_site()` was deprecated and not
left with a full, working replacement.

We know that this environment relies on a single multisite network
and can populate a network object accordingly. If for some reason
multiple networks are used in the future, check to make sure the
populated site object matches the network as expected.