releases: Fix broken stackpath image
The jquery.org domain is having issues serving anything other than
the redirect at "/". The main impact noticable is the broken image
on the releases viewer. Fix that to use a relative reference instead,
since this directory is available from all content domains, and it's
generally preferred to fetch it from the local domain anyway.