events.jquery.org: Remove theme, cross-site integration, and exclusive plugin
* Remove cross-site file inclusion for `$events` variable in mu-plugins

This is in preparation for the simpler standalone WordPress instances
that power the ~20 jQuery documentation sites on new hardware, where
the server will not run as a single WordPress MU instance, and thus
would not have the knowledge of where the other sites are located
on disk or able to include files from it.

The events variable has been empty in practice for a long time.
When needed, we do still sometimes promote events, but we use the
banners for that instead.

* Remove theme

The domain has been replaced with a redirect to openjsf.org in Puppet
for a few years now.

* Remove plugins/allow-numeric-stubs/

Not used by any other sites.