releases.jquery.com: fix broken powered by img
This hardcoded path works fine on the (old) production server,
but on staging, locally, and on the (new) production server this
is a 404 Not Found.

Reference this in a way that lets WordPress control where this
URL is in actuality. This matches the way other images and stylesheets
are referenced throughout the repo, this was the only case of a
hardcoded path that I could find.