mu-plugins/jquery-filters: Remove blog_id condition
In standalone mode, this isn't set.

Follows-up 863a48ed7c61, which added it to support "switch_to_blog".
I don't know what that is exactly, but I know this code can't be having
a useful effect because the surrounding code changed since then.

It now uses JQUERY_LIVE_SITE which can only be set once, so at most
this conditional can change it from applying to not applying overrides.
Maybe there's a context during sunrise.php where this matters, but we
are not creating new sites in the legacy mutlisite setup.