jquery: fix "current" menu marking for local development
The menu functions in this file all use HTTPS, and we this function
compares the declared URL without any filtering, normalization, or
expansion relativel to the current environment or configuration.

As such, this needs to use the same. Note that it was already using
JQUERY_LIVE_SITE which is e.g. "api.jquery.com" even in staging
or locally when browsing via the "local.api.jquery.com" hostname.

Without this, the "Download" page for example is not marked as current
and likewise the "API Documentation" link is not marked as current.