All: avoid gap for broadcast element in site header on mobile
We have a CSS rule for #broadcast:empty that hides the element,
however, the height and margins are applied to the `<aside>` element,
which is the parent element.

Merge these two elements so that there isn't an extra gap.

On desktop, this isn't an issue, because the aside is floated to the
right and is less than than the logo (floated left) and thus
effectively already absent from the layout for all intents and
purposes.