All: Improve logo alignment in site header
Over the years, the addition and subsequent removal of the aside/broadcast
banner to the non-blog theme, has created an unintended drift in where
the logo is aligned in the site header.

The logo was flushed to the left edge of the screen with seemingly
large margin gap above it. Fix this to align with how it renders
on the blog.

Ref https://github.com/jquery/infrastructure-puppet/issues/17