All: move `responsive.html` to `themes/jquery/responsive.html`
Back when the WP-MU setup was used on the old servers, this directory
was mounted at /jquery-wp-content. This is no longer the case on the
new simpler and standalone WordPress installs. To keep using this
utility, expose it via the themes folder instead.

While at, update the URL to be absolute path, so that it naturally
embeds the current site and thus works for all sites, stage/prod,
api, non-api, jquery, jqueryui, etc.