All: Add support for `banner-secondary` to regular Markdown pages
This is currently limited to pages that use page-fullwidth.php.
The default page.php, which includes a sidebar, does not support it,
which means that pages such as those on brand.jquery.org that currently
place banner-secondary into their HTML content, are not able to
switch to Markdown unless they keep hardcoding the HTML, or switch to
a sidebar-less template.