Blogs: Align first article box on homepage with first sidebar heading
This has been misaligned this way for a long time (not a recent regression),
but seems easy to fix. Basically, the margin on `hentry` makes sense
in almost every context: single articles (there's a page title with
horizontal rule before it), and category/tag/search etc all have a
similar page title and horizontal rule.

The exception is the homepage where there is no explicit inline page
title, and instead the first element is the first article box, in which
case its margin is not separating it from anything else but from the
place where it should start.